requirements.txt
python-dotenv
langchain-openai
langchain-core
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
openai
requests
click
PyMuPDF
